+++ modules/tinymce/src/plugins/link/test/ts/browser/SelectedImageFigureLinkTest.ts
@@ -1,6 +1,6 @@
-import { UiFinder, Waiter } from '@ephox/agar';
+import { UiFinder } from '@ephox/agar';
 import { describe, it, before, after } from '@ephox/bedrock-client';
-import { SugarBody } from '@ephox/sugar';
+import { Class, SugarBody } from '@ephox/sugar';
 import { TinyHooks, TinySelections } from '@ephox/wrap-mcagar';
 
 import Editor from 'tinymce/core/api/Editor';
@@ -31,9 +31,9 @@ describe('browser.tinymce.plugins.link.SelectedImageFigureTest', () => {
       </figure>
 `);
     TinySelections.select(editor, 'figure', []);
-    await Waiter.pWait(100);
-    UiFinder.exists(SugarBody.body(), '[title="Insert/edit link"]');
-    UiFinder.notExists(SugarBody.body(), `[title="Insert/edit link"].tox-tbtn--enabled`);
+    await UiFinder.pWaitForState('link toolbar button is NOT active', SugarBody.body(), '[title="Insert/edit link"]', (elm) => {
+      return !Class.has(elm, 'tox-tbtn--enabled');
+    });
   });
 
   it('TINY-8832: link button should be highlighted when there is a link in a figure element', async () => {
@@ -44,8 +44,9 @@ describe('browser.tinymce.plugins.link.SelectedImageFigureTest', () => {
       </figure>
 `);
     TinySelections.select(editor, 'figure', []);
-    await Waiter.pWait(100);
-    UiFinder.exists(SugarBody.body(), `[title="Insert/edit link"].tox-tbtn--enabled`);
+    await UiFinder.pWaitForState('link toolbar button is active', SugarBody.body(), '[title="Insert/edit link"]', (elm) => {
+      return Class.has(elm, 'tox-tbtn--enabled');
+    });
   });
 
 });